python/psy-core-hermes/pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "psy-core-hermes"
version = "0.1.4"
description = "Tamper-evident memory audit log for Hermes Agent. Hash-chained, HMAC-sealed. Plain plugin observer that forwards every memory mutation to the psy-core audit chain."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "jethros-projects" }]
keywords = [
  "audit",
  "audit-log",
  "tamper-evident",
  "hermes",
  "hermes-agent",
  "memory",
  "ai-agent",
  "observability",
  "governance",
  "psy-core",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX",
  "Operating System :: MacOS",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries",
  "Topic :: System :: Logging",
]
dependencies = [
  "pydantic>=2,<3",
  "pyyaml>=6,<7",
  "watchdog>=4,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.23,<1",
  "ruff>=0.5",
  "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/jethros-projects/psy-core"
Source = "https://github.com/jethros-projects/psy-core/tree/main/python/psy-core-hermes"
Issues = "https://github.com/jethros-projects/psy-core/issues"
Changelog = "https://github.com/jethros-projects/psy-core/blob/main/CHANGELOG.md"

[project.entry-points."hermes_agent.plugins"]
# Hermes's PluginManager loader does `ep.load()` and then looks up
# `register` on the returned object. That means the entry-point value
# must point at the MODULE (so `ep.load()` returns the module, and
# `getattr(module, "register")` resolves to our function), not at the
# function attribute. Verified against hermes-agent v0.11.0 and v0.12.0 at
# hermes_cli/plugins.py: `_load_entrypoint_module` returns `ep.load()`,
# `_load_plugin` calls `getattr(module, "register", None)`.
psy = "psy_core.hermes.register"

[project.scripts]
psy-core-hermes = "psy_core.hermes.cli:main"

[tool.hatch.build.targets.wheel]
# Ship the entire psy_core namespace (only psy_core/hermes/ for now).
# Other psy_core_<adapter> distributions can contribute to the same
# namespace via PEP 420 — there is intentionally no __init__.py at
# src/psy_core/ for that reason.
packages = ["src/psy_core"]

[tool.hatch.build.targets.sdist]
include = [
  "/src/psy_core",
  "/README.md",
  "/pyproject.toml",
  "/tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "PL", "RUF"]
# PLR0913/PLR2004/PLR0915: too-many-args, magic-value, too-many-statements; not useful for a small library.
# PLC0415: lazy import. We deliberately defer expensive imports (yaml, watchdog) and import at use sites
# to keep startup lean and to break circular-import chains in hooks↔watcher.
# RUF059: unpacked-variable. We sometimes destructure all four return values for readability even when
# only some are used in a particular test.
# B017: assert blind exception. Tests assert "this raises" without caring about the type when the lib
# legitimately raises pydantic.ValidationError but we don't want to pin to that specific type.
# E501: line-length is enforced via formatter, not the linter, in this scope.
# PLW2901: loop var overwrite by assignment is a legitimate pattern for `for line in stdin: line = line.strip()`.
ignore = [
  "PLR0913",
  "PLR2004",
  "PLR0915",
  "PLC0415",
  "RUF059",
  "B017",
  "E501",
  "PLW2901",
  # Tests intentionally call subprocess.run without check= so we can
  # examine returncode + stderr in the assertion. Adding check=True would
  # turn a failure into a CalledProcessError and lose the diagnostic.
  "PLW1510",
  # Comments use `×` (multiplication sign) for matrices like
  # `action × target`. The look-alike check is noise here.
  "RUF003",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["error"]

[tool.mypy]
python_version = "3.11"
strict = true
disallow_untyped_decorators = false
ignore_missing_imports = true
